+++ iotests/aioRequests.c
@@ -144,7 +144,7 @@ double aioMultiplePositions(const int fd,
     if (ret > 0) {
 
       // verify it's all ok
-      for (size_t j = 0; j < ret; j++) {
+      for (int j = 0; j < ret; j++) {
 	struct iocb *my_iocb = events[j].obj;
 	if (events[j].res != len) { // if return of bytes written or read
 	  fprintf(stderr,"%ld %s %s\n", events[j].res, strerror(events[j].res2), (char*) my_iocb->u.c.buf);
@@ -211,7 +211,7 @@ int aioVerifyWrites(const char *path,
       }
       bytesRead = read(fd, buffer, len);
       buffer[len] = 0;
-      if (bytesRead != len) {
+      if ((size_t)bytesRead != len) {
 	fprintf(stderr,"[%zd/%zd] verify read truncated bytesRead %d instead of len=%zd\n", i, pos, bytesRead, len);
 	errors++;
       } else {
